Update fvmbook scripts to work with Clawpack 5.5.0.
