--- fvmbook/chap11/burgers/Makefile
@@ -1,8 +1,0 @@
-PYCLAWMAKE = $(PYCLAW)/Makefile.common
-
-RP_SOURCE = $(RIEMANN)/src/rp1_burgers.f90
-
-all: classic1.so sharpclaw1.so
-
-include $(PYCLAWMAKE)
-
