--- fvmbook/chap3/acousimple/Makefile
@@ -1,7 +1,0 @@
-PYCLAWMAKE = $(PYCLAW)/Makefile.common
-
-RP_SOURCE =  $(RIEMANN)/src/rp1_acoustics.f
-
-all: classic1.so sharpclaw1.so 
-
-include $(PYCLAWMAKE)
