--- fvmbook/chap6/Makefile
@@ -1,8 +1,0 @@
-PYCLAWMAKE = $(PYCLAW)/Makefile.common
-
-RP_SOURCE = $(RIEMANN)/src/rp1_advection.f
-
-all: classic1.so sharpclaw1.so classicimplicit1.so
-
-include $(PYCLAWMAKE)
-
