+++ fvmbook/chap6/advection.py
@@ -7,45 +7,37 @@ if IC=='gauss_square':
 elif IC=='wavepacket':
     beta=100.; x0=0.5; mx=100
 
-def fig_61_62_63(kernel_language='Python',iplot=False,htmlplot=False,solver_type='classic',outdir='./_output'):
+def fig_61_62_63(solver_order=2, limiters=0):
     """
     Compare several methods for advecting a Gaussian and square wave.
 
     The settings coded here are for Figure 6.1(a).
     For Figure 6.1(b), set solver.order=2.
-    For Figure 6.2(a), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.minmod
-    For Figure 6.2(b), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.superbee
-    For Figure 6.2(c), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.MC
+    For Figure 6.2(a), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.minmod (1)
+    For Figure 6.2(b), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.superbee (2)
+    For Figure 6.2(c), set solver.order=2 and solver.limiters = pyclaw.limiters.tvd.MC (4)
 
     For Figure 6.3, set IC='wavepacket' and other options as appropriate.
     """
     import numpy as np
     from clawpack import pyclaw
+    from clawpack import riemann
 
-    if solver_type=='sharpclaw':
-        solver = pyclaw.SharpClawSolver1D()
-    else:
-        solver = pyclaw.ClawSolver1D()
-
-    solver.kernel_language = kernel_language
-    from clawpack.riemann import rp_advection
-    solver.num_waves = rp_advection.num_waves
-    if solver.kernel_language=='Python': 
-        solver.rp = rp_advection.rp_advection_1d
+    solver = pyclaw.ClawSolver1D(riemann.advection_1D)
 
     solver.bc_lower[0] = 2
     solver.bc_upper[0] = 2
-    solver.limiters = 0
-    solver.order = 1
+    solver.limiters = limiters
+    solver.order = solver_order
     solver.cfl_desired = 0.8
 
-    x = pyclaw.Dimension('x',0.0,1.0,mx)
-    grid = pyclaw.Grid(x)
+    x = pyclaw.Dimension(0.0,1.0,mx,name='x')
+    domain = pyclaw.Domain(x)
     num_eqn = 1
-    state = pyclaw.State(grid,num_eqn)
-    state.problem_data['u']=1.
+    state = pyclaw.State(domain, num_eqn)
+    state.problem_data['u'] = 1.
 
-    xc=grid.x.center
+    xc = domain.grid.x.centers
     if IC=='gauss_square':
         state.q[0,:] = np.exp(-beta * (xc-x0)**2) + (xc>0.6)*(xc<0.8)
     elif IC=='wavepacket':
@@ -54,15 +46,12 @@ def fig_61_62_63(kernel_language='Python',iplot=False,htmlplot=False,solver_type
         raise Exception('Unrecognized initial condition specification.')
 
     claw = pyclaw.Controller()
-    claw.solution = pyclaw.Solution(state)
+    claw.solution = pyclaw.Solution(state, domain)
     claw.solver = solver
-    claw.outdir = outdir
 
-    claw.tfinal =10.0
-    status = claw.run()
+    claw.tfinal = 10.0
+    return claw
 
-    if htmlplot:  pyclaw.plot.html_plot(outdir=outdir)
-    if iplot:     pyclaw.plot.interactive_plot(outdir=outdir)
 
 if __name__=="__main__":
     from clawpack.pyclaw.util import run_app_from_main
